# requisitos/admin/ej-admin-openapi.yaml
openapi: 3.0.3
info:
  title: Elastic Journey Admin Portal API
  description: >
    API do Elastic Journey Admin Portal para gestão de produtos (e dos tipos de
    canal que cada um habilita), jornadas multicanal, fluxos, formulários,
    execuções e publicações. A publicação aciona uma API externa mockada na
    versão 1.0.0; este contrato descreve somente a API do Admin Portal.
  version: 1.0.0
servers:
  - url: https://admin.elasticjourney.com/api/v1
    description: Production
  - url: https://admin-hml.elasticjourney.com/api/v1
    description: Homologation

# Na versão 1.0.0, a autenticação utiliza um provedor externo mockado. O acesso inicial
# usa o usuário admin, senha admin e papel ADMIN; as operações devem preservar
# a identidade autenticada e aplicar os papéis ADMIN, EDITOR e VIEWER.
security:
  - bearerAuth: []

# Observabilidade (FT-10): toda requisição aceita o header opcional
# X-Correlation-Id; quando ausente, o servidor gera um valor. O mesmo valor é
# devolvido no header de resposta X-Correlation-Id e correlaciona, nos logs
# técnicos do backend, a entrada/saída da requisição às transações de
# persistência por ela disparadas. Cross-cutting, omitido dos schemas de
# request/response de cada operação abaixo para não repetir o mesmo header em
# cada endpoint.
...
paths:
...
  /journeys/{journeyId}/flow:
    get:
      tags: [Flows]
      summary: Get journey flow
      operationId: getFlow
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Flow found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
    put:
      tags: [Flows]
      summary: Update journey flow
      description: >
        Replaces the complete flow. The backend returns 422 unless the payload
        contains exactly one initial element (START or MESSAGE_START_EVENT),
        exactly one END and every node belongs to a continuous path between them.
      operationId: updateFlow
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flow'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Flow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
  /journeys/{journeyId}/flow/nodes/{nodeId}/connector-test:
    post:
      tags: [Flows]
      summary: Test a REST connector configuration (REQ-03.10.001)
      description: >
        Executes a REST call server-side with the currently configured method/url/headers/body
        (not the saved flow) and returns the raw response, so the editor can help write
        outputMapping jsonPath rules. Never called directly from the browser (REQ-03.10.002).
        The backend rejects URLs resolving to private, loopback or reserved addresses
        (REQ-03.10.003) and applies a short timeout / response size limit (REQ-03.10.004).
        Not recorded as a domain audit transaction.
      operationId: testConnector
      parameters:
        - $ref: '#/components/parameters/JourneyId'
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectorTestRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: URL resolves to a private/loopback/reserved address, or the upstream call failed/timed out.
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Raw response from the tested call
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorTestResponse'

  # Component Registry (US-04.07, novo em 2026-09-05) — catálogo de componentes SDUI que o Form
  # Builder pode usar numa tela (ver SduiNode/ComponentDefinition acima).
  /component-registry:
    get:
      tags: [ComponentRegistry]
      summary: List all components in the catalog (REQ-04.07.009/010)
      operationId: listComponentDefinitions
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: All components, including REMOVED ones (kept for historical screens).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComponentDefinition'
    post:
      tags: [ComponentRegistry]
      summary: Create a new component (REQ-04.07.001/002, ADMIN only)
      operationId: createComponentDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentDefinition'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A component with the same type+version already exists.
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentDefinition'
  /component-registry/{id}:
    put:
      tags: [ComponentRegistry]
      summary: Update a component (ADMIN only)
      operationId: updateComponentDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentDefinition'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentDefinition'
    delete:
      tags: [ComponentRegistry]
      summary: Mark a component as REMOVED (REQ-04.07.011, ADMIN only) — never deletes the row.
      operationId: deleteComponentDefinition
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '204':
          description: Marked REMOVED
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: mock-session
  responses:
    BadRequest:
      description: Malformed request or invalid parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or unauthenticated identity; reserved for future use
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated identity is not allowed to perform the operation; reserved for future use
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Request conflicts with the current resource state or a uniqueness constraint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: Request is syntactically valid but violates a functional rule
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Unexpected internal failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
...
    JourneyId:
      name: journeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        code:
          type: string
          description: Stable machine-readable error identifier
        message:
          type: string
          description: Human-readable error description
        path:
          type: string
          description: Request path that produced the error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
...
    Flow:
      type: object
      description: >
        Persisted as a single JSONB document per journey (flow, journeyId,
        name, nodes, connections, annotations). flowId, FlowNode.nodeId and
        FlowConnection.connectionId are the identifiers later embedded
        verbatim as BPMN element ids by the runtime when the flow is
        translated to a runtime engine process definition; XML ID/NCName forbids an
        identifier starting with a digit, which a raw UUID cannot guarantee.
        These three identifiers are therefore generated as a fixed prefix
        + UUID (Process_<uuid>, Node_<uuid>, Flow_<uuid> respectively), never
        a bare UUID. All other entity identifiers in this API (productId,
        journeyId, formId, ...) never appear inside generated BPMN
        XML and remain plain UUIDs. annotations are canvas-only notes, never
        translated to BPMN and never embedded with a runtime-facing id.
      required: [flowId, journeyId, name, nodes, connections, annotations]
      properties:
        flowId:
          type: string
          example: Process_f47ac10b-58cc-4372-a567-0e02b2c3d479
        journeyId:
          type: string
          format: uuid
        name:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/FlowNode'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/FlowConnection'
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/FlowAnnotation'
    FlowNode:
      type: object
      required: [nodeId, nodeType, name]
      properties:
        nodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
        nodeType:
          type: string
          enum: [START, END, USER_TASK, SERVICE_TASK, RECEIVE_TASK, MESSAGE_START_EVENT]
        name:
          type: string
        description:
          type: string
        positionX:
          type: integer
        positionY:
          type: integer
        userTaskConfig:
          $ref: '#/components/schemas/UserTaskConfig'
        connectorConfig:
          $ref: '#/components/schemas/IntegrationTaskConfig'

    # Nota de revisão (2026-09-05): embeddedScreen (lista de FormField) substituído por
    # embeddedScreenRoot (SduiNode único) — catálogo SDUI corporativo v1, ver ej-admin-requisitos.md
    # FT-04. A árvore editada aqui é a mesma árvore publicada, sem etapa de compilação.
    UserTaskConfig:
      type: object
      properties:
        embeddedScreenRoot:
          description: >
            Raiz da árvore de tela (SduiNode) desenhada diretamente no nó — ausente quando a
            User Task não tem tela (nesse caso, messageText é usado). Nunca coexiste com
            sentido junto de messageText. Sempre um único nó do tipo ui.screen.
          $ref: '#/components/schemas/SduiNode'
          nullable: true
        messageText:
          type: string
    IntegrationTaskConfig:
      type: object
      description: >
        Exposed as FlowNode.connectorConfig — not a standalone resource:
        connectorType/config/credentialRef are attributes of the FlowNode
        item itself (Flow.nodes, JSONB array). There is no connectorEnabled
        per node: whether a connector type is available at all is a
        catalog-level filter on the type list offered to the user, not a
        stored field. `config` is a free-form map, not a discriminated
        union — RestConnectorConfig/KafkaConnectorConfig below document its
        real shape per connectorType for reference, they are not separate
        schema types on the wire.
      required: [connectorType, config]
      properties:
        connectorType:
          type: string
          enum: [REST, KAFKA, EVENT_HUBS, SERVICE_BUS]
        config:
          type: object
          additionalProperties: true
          description: >
            REST: see RestConnectorConfig. KAFKA/EVENT_HUBS/SERVICE_BUS: see
            KafkaConnectorConfig. outputMapping lives inside this map for
            both — there is no separate top-level outputMapping field, and
            no inputMapping field at all (text fields like url/headers/body/
            payload just reference prior-step variables inline via
            `{{name}}`, REQ-03.09.012).
        credentialRef:
          type: string
          maxLength: 200
          description: Runtime-resolved credential reference; secret values are not stored.
...
    ConnectorTestRequest:
      type: object
      required: [method, url]
      description: REST-only (REQ-03.10.001); values as currently configured in the editor, before saving the flow.
      properties:
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        url:
          type: string
          format: uri
        headers:
          type: object
          additionalProperties: true
        body:
          type: object
          additionalProperties: true
        sampleVariables:
          type: object
          additionalProperties:
            type: string
          description: Example values for `{{name}}` references present in method/url/headers/body, supplied manually for the test call (REQ-03.09.012, REQ-03.10.005).
    ConnectorTestResponse:
      type: object
      required: [status, headers, body]
      properties:
        status:
          type: integer
        headers:
          type: object
          additionalProperties: true
        body:
          description: Raw response body, used to help write outputMapping jsonPath rules.
    FlowConnection:
      type: object
      required: [connectionId, sourceNodeId, targetNodeId]
      properties:
        connectionId:
          type: string
          example: Flow_f47ac10b-58cc-4372-a567-0e02b2c3d479
        sourceNodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
        targetNodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
    FlowAnnotation:
      type: object
      description: >
        Free-floating note on the flow canvas, documentation-only — never
        validated by FlowValidator, never translated to BPMN. id is a plain
        Annotation_<uuid> by convention, never embedded in generated BPMN
        XML, so it carries no NCName constraint.
      required: [id, text, positionX, positionY]
      properties:
        id:
          type: string
          example: Annotation_f47ac10b-58cc-4372-a567-0e02b2c3d479
        text:
          type: string
        positionX:
          type: integer
        positionY:
          type: integer
        linkedNodeIds:
          type: array
          items:
            type: string
            example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
          description: FlowNode ids in the same flow this note is visually tied to (dashed line in the editor). Informational only.

    # FormField (modelo de campo plano do catálogo SDUI anterior) removido em 2026-09-05, junto com
    # US-04.02 — ver ej-admin-requisitos.md FT-04. Substituído por SduiNode (árvore recursiva) +
    # ComponentDefinition (Component Registry, schema mais abaixo).
    SduiNode:
      type: object
      description: >
        Nó da árvore de tela do catálogo SDUI corporativo v1 — substitui por completo o antigo
        FormField (lista plana) e a tupla [tag, props, children]. A raiz de uma tela é sempre um
        único SduiNode do tipo "ui.screen"; type+version devem existir no Component Registry.
      required: [id, type, version]
      properties:
        id:
          type: string
          description: >
            Identificador único dentro da tela — para um componente que coleta valor, também
            usado como sufixo do binding de leitura-e-escrita no namespace form (REQ-04.10.005).
        type:
          type: string
          example: ui.textInput
          description: Tipo do componente, correspondente a ComponentDefinition.type no Component Registry.
        version:
          type: string
          example: "1.0"
        props:
          type: object
          additionalProperties: true
          description: Configuração estática do componente, conforme o propsSchema declarado pelo ComponentDefinition correspondente.
        bindings:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SduiBinding'
          description: 'Hoje só a chave "value" é usada (REQ-04.10.001/002).'
        events:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SduiEvent'
          description: Chave = nome do evento (ex. onPress, onChange), restrito aos eventos que o ComponentDefinition declara.
        visibility:
          $ref: '#/components/schemas/SduiVisibility'
          nullable: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/SduiNode'
          description: Presente só quando o componente aceita filhos (ComponentDefinition.allowsChildren); profundidade não limitada.
    SduiBinding:
      type: object
      required: [path, mode]
      properties:
        path:
          type: string
          example: form.customerEmail
          description: 'namespace.resto — namespace ∈ {form, data, session, route, computed} (REQ-04.10.001).'
        mode:
          type: string
          enum: [oneWay, twoWay]
    SduiEvent:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [action.submit, action.navigate, action.openUrl, action.setValue, action.track, action.dismiss]
        params:
          type: object
          additionalProperties: true
    SduiVisibility:
      type: object
      required: [rule, path]
      properties:
        rule:
          type: string
          enum: [equals, notEquals, in, notIn]
          description: >
            equals/notEquals compare against a scalar value. in/notIn (added 2026-09-06)
            compare against a list value — used mainly for path session.channel, to
            condition a component on a subset of the journey's channel types.
        path:
          type: string
        value: {}

    # Component Registry (US-04.07) — fonte de verdade do que o Form Builder pode compor numa tela.
    # Persistido em tabela (component_definition), não mais uma lista fixa em código.
    ComponentDefinition:
      type: object
      required: [type, version, status, level, category, allowsChildren]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        type:
          type: string
          example: ui.textInput
        version:
          type: string
          example: "1.0"
        status:
          type: string
          enum: [EXPERIMENTAL, STABLE, DEPRECATED, REMOVED]
        level:
          type: integer
          minimum: 0
          maximum: 4
        category:
          type: string
          enum: [CONTENT, LAYOUT, INPUT, ACTION, FEEDBACK]
        allowsChildren:
          type: boolean
        allowedChildTypes:
          type: array
          items:
            type: string
          description: Vazio = qualquer tipo de filho permitido.
        propsSchema:
          type: array
          items:
            $ref: '#/components/schemas/PropDescriptor'
        events:
          type: array
          items:
            type: string
        supportedTargets:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/TargetSupport'
          description: 'Chaves: react.web, react.mobile, flutter.web, flutter.mobile.'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    PropDescriptor:
      type: object
      required: [name, kind, required]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [TEXT, NUMBER, BOOLEAN, ENUM, TOKEN, OPTIONS_LIST, VALIDATION_LIST]
        required:
          type: boolean
        defaultValue: {}
        tokenGroup:
          type: string
          nullable: true
          description: Só relevante quando kind = TOKEN (ex. color, spacing — seção 10 do contrato).
        enumValues:
          type: array
          items:
            type: string
          nullable: true
    TargetSupport:
      type: object
      required: [status, minRendererVersion]
      properties:
        status:
          type: string
          enum: [SUPPORTED, PLANNED, UNSUPPORTED]
        minRendererVersion:
          type: string
          example: 1.0.0

    # Nota de revisão (2026-09-05): embeddedScreenSdui (árvore compilada) substituído por
    # embeddedScreenRoot em SnapshotFlowNode — a mesma árvore de SduiNode editada no Form
    # Builder, sem etapa de compilação separada. Nota de 2026-08-24 mantida abaixo por histórico.
    #
    # Nota de revisão (2026-08-24): "forms"/PublicationSnapshotForm removidos, "versionNumber"
    # adicionado, e flowNodes passou a referenciar SnapshotFlowNode (não mais FlowNode) nesta
    # revisão — a Runtime Engine só suporta um conjunto básico de tipos de campo nativos (~5-6),
    # inviabilizando manter a User Task associada a um formulário do catálogo por formId; a tela
    # passou a ser desenhada diretamente no nó, e o snapshot enviado ao runtime não carrega mais
    # uma lista de formulários — só a tela já compilada (embeddedScreenSdui) de cada nó.
...